Add "refund insufficent funds" config option
Add the option to not refund buyer when the amount of krist paid is less than the amount needed to buy one of the desired product. This is a very evil option, but is the last refund type that doesn't either have a hook or config option to suppress it.